+++ NEWS.md
@@ -54,5 +54,5 @@ function to find the parameters. Equivalent changes has made to fitbeta and
 fitbeta_ml
 
 # convdistr 1.4
-* new_DIRICHLET now accept to have a columm with zero probability. If only
-one column have probability, returns 1 in that colum and zero elsewhere
+* new_DIRICHLET now accept to have a column with zero probability. If only
+one column have probability, returns 1 in that column and zero elsewhere